requirements.txt
django==5.0.6
djangorestframework==3.15.1
django-debug-toolbar==4.4.2
django-import-export[xlsx]==4.1.1
djoser==2.2.3
environs[django]==11.0.0
gunicorn==20.1.0
mssql-django==1.5
psycopg2-binary==2.9.9
